renaming and testing of basic workflow
