+++ msg/Navigation.msg
@@ -1,0 +1,4 @@
+float64 depth
+float64 heading
+float64 vx
+float64 vy
